requirements.txt
aiohttp==3.8.4
pillow==9.4.0
